Yield NodeDriver thread to avoid starvation
Empirically, the run-node thread starves the communication
thread when acquiring the Node. This causes transmission to be
delayed. Now, when a request from Dispatch is received, a flag is set,
and the run-node thread will yield its CPU time slice before acquiring
the node. In theory, this will allow the communication thread to acquire
the node. Ultimately, I expect this to be at best a temporary or partial
solution, as I believe much work can be done in terms of resource
sharing between threads in the nodes.